Project detail error solve

+++ src/main/java/com/bot/projects/entity/Projects.java
@@ -6,6 +6,7 @@ import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -72,7 +73,7 @@ public class Projects {
     Date updatedOn;
     @Transient
     @JsonProperty("CTC")
-    float cTC;
+    BigDecimal cTC;
     @Transient
     @JsonProperty("EmployeeId")
     long employeeId;
